fix: move speech recognition to hooks
